Displaying matches per stage.
