+++ app/src/main/java/com/adammcneilly/pocketleague/ui/EventStageDetailContent.kt
@@ -1,10 +1,11 @@
 package com.adammcneilly.pocketleague.ui
 
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Text
+import androidx.compose.material3.Divider
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -23,9 +24,13 @@ fun EventStageDetailContent(
             .fillMaxSize(),
     ) {
 
-        Text(
-            text = "Event Stage Detail"
-        )
+        Column {
+            viewState.matches.forEach { match ->
+                MatchListItem(displayModel = match)
+
+                Divider()
+            }
+        }
 
         if (viewState.showLoading) {
             CircularProgressIndicator(
